Resolved camera intent crashing app.
I forgot to change the target SDK inthe build.gradle file.
Basically, and I don't think Cai really explained why we did this unless I just missed it, in SDK 23 or newer Google has changed how we have to check permissions.

<22: Just have it in the manifest, and it will automatically work.
>23: You have to have it in the manifest to tell the app it NEEDS this, but you also have to include lines of code in the specific spot those permissions are going to be used to first confirm that they have been given by the user for the app to continue.

In my case, because I didn't have those lines of code, it simply attempted to continue regardless of whether it was allowed to create a new file or not.

* In short, setting the target SDK resolved the issue. Will need to learn the correct, newer way at some point.